Bug 775835 - Update patch name and description with bug number (if requested) even when attaching to an existing bug. r=jdm
With my own workflow, this turns out to make way more sense. In particular, if you have multiple attachments for a single bug, this is the only thing that makes sense. But even if not, it's nearly always what I want. Since updating is currently non-default anyway and nobody ever requested the previous behavior, I feel comfortable making this change.