Add the reviews / view diff header template.
This adds the big "Reviews" and "Diffs" links at the top of the review
request and diffviewer pages. It also hides the review request details
in the diffviewer.

Still missing are styles for the big buttons at the top. We need to
beautify this quite a bit.

--HG--
extra : convert_revision : 1dddf14d09285b3feb1d2c9a644996a1bc7d16e0