bundlerepo: make compatible with 2.5.4
Mozilla still runs Mercurial 2.5.4 on the server. So, we backport
bundleclone to work with it.

The new code has been tested on all versions between 2.5.4 and 3.0.2 and
appears to work, so all versions are marked as compatible.

--HG--
extra : rebase_source : 169fe3fff4bba0708aa06a4e061b242a548e96f6