testing: support hghave
run-tests.py expects hghave to be in the test's directory. This is
subpar for our use as we'd like a single hghave script. We'll live
with this now. But we should consider a better solution if we need
hghave in multiple directories.

--HG--
extra : histedit_source : 66d3a0dc77cb35f2bf1f8737209c31a64b91fdab