pushlog: add a test for rollback behavior; r=glandium
The pushlog doesn't currently handle transaction rollbacks properly. It
may record changes even if the push/transaction is aborted.

We add a test that verifies the existing behavior so that the behavior
change is positively identified when it is implemented later.

--HG--
extra : rebase_source : a089691a0f12584ed5f40d05181fd0cbf1441ca3