Bug 1067525 - Use the bugzilla cookie login resource in post_reviews if a cookie is provided. r=gps
This makes use of the new bugzilla cookie login API to login to Review
Board. If a userid and cookie value are provided to the post_reviews
function it will ignore any provided username/password and attempt to
login using the cookie.

This introduces some Bugzilla specific code into the pushhooks.py file,
which is unfortunate. Since this is all just a big hack until there is
proper OAuth support in both Review Board and Bugzilla we'll deal with
it and try to avoid allowing any *more* bugzilla specific code into
this piece of the system.

--HG--
extra : rebase_source : 3459897d2c876101215321644217e5fd4df7d774