bzpost: remove old logging of bug activity (bug 1078659); r=glandium
As part of 1e900584cb2d, we added logging code to the bzpost extension
to facilitate testing. This was to work around testing deficiency around
use of mocks.

Now that we run a local BMO instance as part of the tests, we can dump
the state of BMO as part of tests. This is a better method of testing
than mocking because we are actually testing that Bugzilla is updated,
not just that we think we're updating Bugzilla. So, we remove the code
for dumping what comments Bugsy is writing and replace it with dumping
of actual Bugzilla state.

--HG--
extra : rebase_source : fdffdd3b93d37b4dad443d57d63bfeb844b5ce1f
extra : amend_source : 28dda0feec97a5e6ef2d7a091f610dad847ca29a