Bug 1067525 - Allow authenticating with Bugzilla Cookie when anonymous access is disabled. r=mconley
Before this patch if Review Board was set to disallow anonymous access
you would not be able to retrieve the API resource to make a POST
request. We now always allow anonymous requests to the new resource so
that you can use it to authenticate when Review Board is locked down.

--HG--
extra : rebase_source : 097f622680489734e34575ad666a46276a8e6e52