mozext: drop support for Mercurial <3.0
Mercurial has been making significant changes to its internal APIs
around exchange. This makes continued support for <3.0 more and more
difficult. We could continue to support <3.0 with this extension. But
this author believes now is as good a time as any to drop support. So
this patch drops marked compatibiity with everything less than 3.0.