bundleclone: drop support for Mercurial <3.0
The bundleclone extension is experimental. There are enough refactorings
in Mercurial to make continued support of 2.5.4 and other old releases
difficult.

At the time this extension was written, Mozilla was stuck in Mercurial
2.5.4 on the servers and in Firefox automation. The transition to 3.1+
is already in progress, so continued support for 2.5.4 has marginal
value.

Support for Mercurial <3.0 is thus being dropped.