Bug 1085417 - Stop using 'Review' in RB summaries. r=gps
We had been using 'Review for Review ID: <review-id>' as the summaries
in Review Board. This uses the wrong terminology ('Review') to refer to
a Review Request.

We now set the summary to just be '<review-id>' and set the bugzilla
attachment name and description to be 'MozReview Request: <review-id>'.

--HG--
extra : histedit_source : 1e40a5738920ed36c043aaa94a19dbd1f283ffe8