rbmozui: update reviewers when we have an autocomplete result (bug 1087491); r=mconley
Before, we needed to submit the reviewer autocomplete <input> <form> by
pressing enter to trigger addition of a reviewer.

With this patch, we add a reviewer as soon as the autocomplete box says
we have a result. Selecting a user with a mouse click or pressing tab
will now instantly add a reviewer. The user experience is much better.

If a user clicks outside the autocomplete element, it is possible for
an unsubmitted value to remain in the element and this value will not be
recorded. This edge case should be improved in a future patch.

--HG--
extra : rebase_source : 53692290e66beb68aeb8812cf0f756496408e7f5
extra : amend_source : 7fe65b67e26628fe13a6511ebc555c268d3b28f4