docker: properly use per-test state files
The previous logic around Docker state file use accidentally had all
tests using the same state file. This resulted in race conditions
updating the file/state. This meant that some containers were not being
recorded in the state file and when it came time to stop the containers,
no record of them could be found and they kept running. This likely
resulted in out of memory errors when running lots of tests, such as
an --all-versions test run. This likely caused the Jenkins jobs to start
timing out recently.

With proper per-test state files, there are no race conditions and
containers shut down properly at the end of tests. We can thus remove
the redirection in various tests hiding what was formerly inconsistent
output and not suitable for testing. (As it turns out, this redirection
was masking a real bug!)

With expected output in place, it was revealed that some tests didn't
have cluster names agreeing, causing no containers to shut down. In one
case, we accidentally shut down another test's containers if it was
running!

With all the changes in place, containers now properly shut down at the
end of tests and everything is happy.

There is still a bug around container deletion on AUFS. But that's for
another commit.

--HG--
extra : amend_source : 7b5a669ba76f302509fa5d368325a02ebd54c4e8