mozreview: don't show the commits tab to non-submitters (bug 1047468); r=mconley
Actions on the Commits tab on the main review request interface only
make sense to the submitter of the review request. So, we hide the
Commits tab for non-submitters.

This approach is far from a robust solution. For one, there is no
authorization on the URL, so anyone can type the magic words and still
see the Commits page. However, this isn't an immediate concern.

There is also information on the Commits page that might be useful to
non-submitters. The eventual plan is to move data from the Commits page
into the main review request page. At that time, we can expose the
useful data to the relevant set of users.

--HG--
extra : rebase_source : 1f5d24eb308d649d11f97f09cb2b36dc82cf8e38
extra : amend_source : cf429259ae1bd6e33b0c7a1c4eb6fdbebec13a63