bugzilla: move testing/bugzilla.py to bugzilla
As we focus on making MozReview easier to hack on, we should expose the
tools to interact with the various systems in a cleaner way. We move
testing/bugzilla.py to ./bugzilla to make it easier to consume.

--HG--
rename : testing/bugzilla.py => bugzilla
extra : amend_source : 56a344739c32ea81d8b70f646b14801fc3eff66e