reviewboard: change multiple head detection; r=smacleod
As part of implementing support for limiting reviews to a base
changeset, it was observed that Mercurial (weirdly) attributes each
argument to `hg push -r` as a head during discovery! This was triggering
our existing "prevent multiple headed push" check.

This commit changes that detection mechanism to work with multiple
revisions passed to `hg push`. We don't include a new test because this
will be implicitly tested as part of specifying a base revision for
review.

--HG--
extra : rebase_source : e3b4611f994e6c80aad82946b468fb92ca1bd1f5