reviewboard: support specifying base revision for review (bug 1089685); r=smacleod
People often commit changes to separate features in the same head.
Sometimes there is commit dependency involved. Sometimes it is a user
workflow preference. Either way, the default selection of non-public
ancestors is too greedy if the user wishes to keep the review separate.

This patch introduces the concept of a base review revision. If `hg push
-r` is used and there are multiple revisions to be pushed, we interpret
the bottom-most revision as a request for a base revision to review.

Overloading -r instead of introducing a new argument was chosen because
a) it was simpler b) it mostly makes sense c) it reinforces an existing
Mercurial concept instead of inventing a new one. We want the code
review workflow to fit inside standard Mercurial workflows as much as
possible. People should be familiar with `push -r` because it is a
feature of core Mercurial. When used with `push`, -r means "the
revisions you want to exist on the remote." Applied to a review
repository, it logically makes sense to extend this to "the revisions
you want to review." Since we already enforce the requirement that
pushes to review repositories come from the same head, multiple
revisions specified with -r must be in the same DAG chain. Thus taking
the first and last as the base and tip commits to review seemsa
appropriate. Finally, it is easier to overload -r than to introduce new
arguments. This is a bad reason (UX is a good reason), so it is last.

During ancestor walking, we arbitrary choose to still limit the base
review revision to non-public changesets, even if the user-specified
base is public. We could loosen this requirement and allow people to
review public changesets. However, restricting reviews to non-public
changesets has the benefit that it strongly encourages the use of draft
changesets and keeps the changeset evolution and obsolescence door open
for more people. I think we should be guiding people to use phases
properly and not allowing review of public changesets will help us
uncover workflows where phases aren't being properly applied. This will
help us identify and convert more people to "best practice" workflows.

--HG--
extra : rebase_source : 02366dac6f3717424b213b286a2d824c7e450bca
extra : amend_source : 9c877e76218e692234fc0aed56df1bf3238c896d
extra : histedit_source : a1d1907b3f4b3b56b8fbb2a8f95332218e4b2ca7