reviewboard: clear r? when publishing a review (bug 1096758); r=smacleod
Before this patch, if a non Ship It review were published and the
reviewer had an r? flag for the review request, the r? flag would
linger.

In most situations, publishing a review means "I'm done with the request
for review."

This patch changes behavior so that publishing a non Ship It review will
result in clearing of the r? flag.

There is still a case where a reviewer may wish to retain the r? flag.
We do not currently handle this. We anticipate we'll add a checkbox or
similar UI to the review interface so this intent can be communicated
and result in not touching the r? flag.

--HG--
extra : rebase_source : 8bbf076eea084c50146c5dc654263f53555412b5
extra : amend_source : d88f2a902a2369414a9b5b4a2d950c2ed468af34