pash: ability to change publishing state of repositories (bug 1089385); r=bkero
User repositories are currently publishing (the default behavior). The
role of user repositories dictates that they be non-publishing by
default. So we are planning on mass-changing user repositories to be
non-publishing.

Before this happens, we need a self serve mechanism to allow users to
revert the publishing state of their repositories back to publishing
This patch implements that mechanism.

--HG--
extra : rebase_source : 3d2211461e64dca0f587d081529924005d8d25ff