rbbz: handle unicode in review comments (bug 1096964); r=smacleod
rbbz.diffs.render_comment_plain() was previously failing to handle
unicode in review diff comments. This was because it was mixing and
matching str and unicode types and then joining everything with a str,
causing an implicit encoding using the ASCII codec, which would fail if
non-ascii code points were present in a unicode instance.

We hack around this by ensuring unicode is used for the join.

The function is still not implicit conversion safe. There is still a mix
of str and unicode instances being used. This will fail in Python 3.
This might also fail if there are unicode filenames present. I stopped
short of converting everything to unicode because I'm not sure what
Django is exposing and expecting under the hood. The current patch
unbreaks the most user-visible failure, which is progress.

A test verifying that unicode diff comments work has been added.

--HG--
extra : rebase_source : faca7e48b5dab8c9c0a9436887b4454670b5c3a2