pushlog-feed: don't error when interacting with empty database (bug 1116899)
Previously, attempting to interact with an empty database resulted in a
TypeError attempting to access [0] of a None instance. This patch fixes
that.

Committing and deploying without review since this is impacting
automation and nobody is around to review since it is New Year's Eve :/

--HG--
extra : amend_source : 88b69cf1d433a200a31e2191a04631624d7023a5