hgtemplates: synchronize fileannotate.tmpl; r=bkero
Only minor changes.

As part of testing this, it was noticed that the display of the
annotation page is busted. This is likely the result of missing CSS,
likely due to entries in map not emitting classes. map will be
synchronized in the future.

--HG--
extra : rebase_source : 025b734a796b962e348e8eb7d225ed217c53bfa2