reviewboard: update tests to reflect Bugzilla DB changes; r=smacleod
Unpinning the Bugzilla commit and enabling component watching resulted
in the creation of a 2nd profile in the Bugzilla database. This threw
off MozReview tests that were looking for specific user IDs.

This patch updates the tests and causes the test to pass with a modern
Bugzilla configuration.

--HG--
extra : rebase_source : 1ffa2837d949e2bce9473453515167c1ca0bc9be