bzexport: revert back to bzAPI for resolving the user name (bug 1121764); r=sfink
bzexport always needs to know the Bugzilla username associated with a
request. When we use cookies extracted from a Firefox profile, the
username is not known and must be resolved by querying Bugzilla.

Unfortunately, it appears the REST API as deployed on BMO does not
currently allow us to use cookies to do what we want. If we attempt a
GET /rest/user/%s, Bugzilla fails with "Logged-out users cannot use the
"ids" argument to this function to access any user information."
Bugzilla thinks we are logged out because cookie auth is only allowed on
POST and PUT HTTP methods. When specified on GET, etc requests, cookies
are silently discarded (for the record, we currently send cookies on GET
requests).

Ideally, we'd exchange cookies for a token to use with the REST API.
However, the REST API's login endpoint is only accessible via GET. And,
when you specify cookies via GET, they are ignored. So, there appears to
be no way to obtain a REST token nor perform authenticated GET requests
when all we have are cookies.

bzAPI doesn't have the restriction that cookies can't be used with GET
requests. Instead, you toss some cookie values in the query string and
it happily uses them for authentication.

In this patch, we revert user lookups to go through bzAPI.

The aforementioned "Logged-out users cannot use..." error message was
observed as part of developing this patch and the switch to bzAPI was
confirmed to make it go away.

--HG--
extra : rebase_source : baf06e73d6f03ddfa107bd8f58560e8940f97470
extra : amend_source : b7e85ec8ad5de6e3e9b373f52f6ffc56d54f0ff8