rbmozui / mozreview: Initial work to expose UI and WebAPI for Try Autoland. (Bug 1109218) r=smacleod,brennie.
This is the initial work to write and expose the TryAutolandTriggerResource on the MozReview
extension, and the very early work to expose the UI for hitting that API endpoint in RBMozUI.

Note that RBMozUI is going to be merged into MozReview, so that's what's informing us putting
the new resource and models into MozReview for now.

When the endpoint is hit, if the user requesting the Try Build is the review request submitter,
we communicate with the Try Autoland service via a configurable URL, get a job ID back from
Try Autoland, and add it to a custom table. Try Autoland will then need to hit another web
API endpoint to update us when the landing has occurred, but that resource will be developed
in a later commit.

The front-end to hit the API endpoint is still in the very early, work-in-progress stages. It
currently hits the endpoint, but doesn't do much with success or fail results.

Again, this is mostly foundational work that we want to get landed so that other people can
start contributing with relative ease.

--HG--
extra : rebase_source : 451226d2a0ec423eabdc69913c821995e4175675