mozreview: Properly setup the Site domain when starting Review Board (Bug 1125468). r=gps
Review Board uses information stored in the Django Site domain field for
a few things dealing with urls. We now properly set this to
localhost:$HGPORT1 when starting the server. We'll need to modify this
if we stop using localhost as the domain.

--HG--
extra : rebase_source : daa8fac50422caf11c8bca0233de61a0a868418a