testing: ability to replace HGRCPATH in dummyssh
When using dummyssh, the HGRCPATH set from the client's environment is
likely picked up and used. In complex tests, this may not be desirable,
especially if we're testing differences in configurations between client
and server.

We change the behavior of dummyssh to allow for the existence of an
alternate config file (HGSSHHGRCPATH). This allows for clearer separation
between client and server configs and better simulates how things work
over real SSH.

--HG--
extra : rebase_source : d2750dcac0bcbde8fedee6f693ac93ffb8f58440