testing: search PATH for hg binary
Previously, the path to the hg binary was hard-coded. If we are to use
mozreview in the tests, which run with different versions of Mercurial,
we need to use the Mercurial appropriate for the environment. So,
implement PATH searching for "hg."

--HG--
extra : histedit_source : eb814caa10921b9990a6766574cecb6a237312a8