pushlog: handle stripping (bug 1127480); r=ted
Previously, the pushlog extension didn't handle stripping events. If
someone ran `hg strip`, the pushlog database would:

1) Forever contain entries for changesets that no longer exist
2) Possibly contain invalid revision numbers for changesets

The second one is the most problematic. Because there is a unique
database index on the the "rev" column of the "changesets" table,
this would almost certainly result in a constraint error during
the next push.

In the rare situation that we had to strip changesets from production
repositories, this action was almost always followed up with
manual pushlog database muckery to restore order. If this weren't bad
enough, these operations had to be performed on every machine with a
pushlog database. The process was very brittle and prone to failure.
Having the pushlog extension do it in reaction to a Mercurial "event"
is a more robust method.

This patch teachs the pushlog how to react to changeset stripping.
This is referred to as a "destroying" event in Mercurial - we're
"destroying" changeset data from the revlogs.

This patch hooks pushlog into the "destroying" event and teaches it
how to reconcile the database against the current state of the
repository. Strictly speaking, the newly implemented method is generic
and could be run anywhere to reconcile malformed pushlogs. But that use
case is not yet explicitly supported.

In the ideal world, the pushlog would record stripping events in the
database. We could probably maintain a separate log of these. Since
this patch is being developed to aid in recovery of some production
repositories, time is of the essence and reduced scope is important
to mitigate risk. A good follow-up would be to implement this feature.
For now, we write a log event to the repository's logging mechanism. If
blackbox is installed, this will go to blackbox.log. It's better than
nothing.

--HG--
extra : amend_source : d1338ac70ec1d4d20ed8d86cf3b60654452cca37