Bug 1128555 - Propagate Reviewboard BadRequestError to user; r=gps
Currently we're not catching BadRequestError from rbtools which results
in an 500 - Internal Server Error from Mercurial being displayed to the
user should one occur. This patch catches them and formats the Reviewboard
error message for the user. Unfortunately the Reviewboard error messages
are not that informative, but this is still improves the current situation.

--HG--
extra : rebase_source : 0c769b416667a35447a04bc1727bc33736c3b921