hghooks: don't process strips from commit-message hook (bug 1128698); r=bkero
We don't want our hooks to fire when a strip is performed because it can
cause the strip to abort mid operation. If a changeset is already in the
repository, it made it in and the hook doesn't have any business
rejecting it again.

--HG--
extra : rebase_source : 40617510613d563c0181b6f680429e0acd5bf7ae
extra : histedit_source : c354975285adaa26a6ff0660f4de889c6b366713