vcttesting: add mach command to dump reviewboard account profile; r=gps
This patch adds a mach command to dump the reviewboard user profile. We modify
the user profile when creating a user from bugzilla, but we don't currently
have a way of quering that information after the fact when writing tests.

This requires a path to the reviewboard sqlite database. I've made this an
optional argument. If omitted it defaults to the path from the command's
MozReview instance which makes things easier when writing tests.

The make-admin command also uses sqlite directly. I've modified the path
argument there to make it consistent with the new dump-account-profile command.

--HG--
extra : rebase_source : a4ade54f56647fad76f5d0f5f81007f33a2272a6