reviewboard: implement support for remote repository stripping (bug 1134382); r=smacleod
When a code review becomes security sensitive, we want to hide the
review and any of the commits that were sent to it. Since the review
repositories are public, this means we need to delete commits from the
review repository.

This patch implements support for remote stripping.

While the functionality is generic, I've implemented it as part of the
reviewboard server extension. Mercurial is building "obliterate" support
into core and I don't think it is worth our time to implement a
standalone extension that will likely be replaced by a core feature.

We want remote stripping to be a privileged operation. We encode a list
of users that are allowed to perform strip. We take the user from the
USER variable, which pash (our SSH handler) should set for us from the
SSH user.

We purposefully leave the client in control of which changesets get
stripped. We can add support for stripping from parsed bug numbers if we
need it.

--HG--
extra : rebase_source : c0e64348b46c9670568d5279a8409ce2e7afd6b6
extra : amend_source : 58cb62309d512f0ae082d6f641f008db7af5bab0