qimportbz: default to importing all patches & improve prompt text (bug 1001463)
I doubt that the first patch is always going to be the one and only patch of
interest, so let's default to all, instead of the first. In addition, users
were not aware that they could comma separate the patch ids, so give an example
using that format.

--HG--
extra : rebase_source : 43db45d5f580e5e1f8c902e459617119cd3f3f09