qimportbz: re-prompt if the patch choice wasn't valid (bug 1001463)
Previously if one or more of the patch choices were invalid, an error would be
output but the patch selection would continue with only a partial list. Now, if
one of the choices is not valid, the user is re-prompted for the selection.

--HG--
extra : rebase_source : 290bd95212efbb585a1b9e0c168f19753867d4a9