testing: make the admin user an admin in Review Board; r=smacleod
Up to this point, the admin user was only an admin user in Bugzilla. We
had to explicitly make the user an admin in Review Board. This was for
historical reasons, as tests were performing actions as the admin user.

Now that all tests are using a regular user account, it is safe to make
the admin user an admin in Review Board. So do it and make hacking on
MozReview easier in the process.

--HG--
extra : rebase_source : f89a201a4438357e01a3a60376bd65cabe5fc57e
extra : amend_source : f01e0c3faaffe9bb9ce732210fe9c962dfd548b7