hgmo: optimize Docker image generation
We use the new feature to limit which files are added to the context. We
also consolidate the number of instructions in the Dockerfiles to make
image building faster.

The images still don't build as quickly as I would like. We may have to
move instructions into a shell script that is executed inside the
container.