qimportbz: avoid filename collisions with the temporary patch file (bug 980986)
Prior to this patch, if an imported patch had a filename that matched the bug
number (and the user hadn't overriden patch_format in their local config), an
error would be shown saying that the patch already existed with the same name,
even when the file did not exist prior to running the import.

This is due to the temporary file (created only for the first patch in the
sequence, later patches don't have one) having the same name as our target
filename. To solve this, we add a suffix to the target filename, which will
allow both this and subsequent imports to succeed without conflicts with the
temporary file.

--HG--
extra : rebase_source : e159eee1c79eb870ccee0e2e4e22b5b016b492aa
extra : histedit_source : d1e1abddadb295a1180d360fde915d73d9c701e4