testing: start an LDAP server as part of MozReview
We want the MozReview test environment to look as similar to production
as possible. To do that, we need to run a Mercurial server that's
similar to production. To do that, we need an LDAP server to
authenticate SSH requests against.

This patch starts a Mozilla flavored LDAP server as part of the
MozReview cluster environment. We currently don't do anything with the
server. But this is a pretty large patch already, so stopping here makes
sense.

--HG--
extra : rebase_source : 3b1558caa1f6d429999bae151ef605f712b6983a