Bug 1055021 - mozreview: Create attachments for all children. r?smacleod
An attachment is now created on Bugzilla for each child.  No attachment is
created for the parent.

Review logic has changed a bit as well.  For children, a ship-it results in an
r+ on the corresponding attachment from the reviewer.  A review with no
ship-it clears any '?' review flag with the requestee being the reviewer, and
any '+' with the setter being the reviewer.

For parents, reviews with ship-its are not allowed, and the corresponding UI
functions have been removed.  A review with no ship-it results in a comment
unassociated with any attachment, and no flag changes.

--HG--
extra : rebase_source : 5083926d1e80201a4e38a506e2f1f1b57c29d78a
extra : histedit_source : 5737da95f153e0935d18b716dd76c157a6ec3427