docker: wait longer when checking if Docker is available
The timeout of 1s was too optimistic, as Docker can get very slow when
it is saturated.

Chances are that if the server isn't available, we'll get a closed port
or unreachable network instead of a timeout anyway. So increasing the
timeout should be mostly harmless.

--HG--
extra : rebase_source : 3bb0d3fa6b488da35fab2b8016aa63505201b551