autoland: add support for transplanting from github to mozreview
This adds support for retrieving patches from github using the github api and
pushing these to a mozreview repository. It is assumed for now that the github
account is a mozilla-gecko fork and the destination repository is a gecko
review repo.

--HG--
extra : rebase_source : 2cc513cd296f49cb68a37b210a2574a9a82d76ad