mozhgweb: establish an extension for hgweb customizations
We have a number of hgweb customizations living in pushlog-legacy.
pushlog-legacy is the wrong location for them.

This extension establishes a new mozhgweb extension with the intent of
holding all the hgweb customizations under one roof. We bootstrap the
extension with the content of buglink.py. We extract tests from
buglink.py into .t tests and we change some formatting to adopt modern
practices.

--HG--
rename : hgext/pushlog-legacy/buglink.py => hgext/mozhgweb/__init__.py