testing: support refreshing hgrb container
With this change, the hgrb container is refreshed as part of `mozreview
refresh`. Currently, the refresh will only synchronize
version-control-tools. This is sufficient for refreshing code running on
the Mercurial server.

As part of this change, the watchman trigger is now defined in terms of
a more complex watchman query expression instead of a single PCRE. This
is much more readable.

--HG--
extra : amend_source : de793d149e999dd82b362af4c21942617dd9eb88