ansible: convert repository re-cloning to a module
The old method of doing re-clones deleted all repos first, then iterated
over them to clone them. This approach was sub-par.

The mirror-pull script takes out a lock when it creates a new repo. If
we removed multiple repositories and someone pushed to one of the
repositories under maintenance before it was cloned, the push would
hang, waiting for a lock.

By rewriting this series of steps as a single module, we reduce the
window of repository locking to only when that specific repository is
under maintenance. A push could still hang waiting for a repository
lock. But there's not much we can do about that without changing the
underlying mirroring architecture.

Also note that we must use an Ansible module for this because Ansible
includes and roles cannot be used with "with_items." Apparently this
feature will be added to Ansible 2, so we should be able to rewrite the
module as a playbook once we upgrade to Ansible 2.

--HG--
extra : rebase_source : 75f8c4c0ba8857fe0ef4e362c656dfdaba51ba8f
extra : histedit_source : 7e1514b1c879dd3c5a4bea73769ff7fcb8baf50c%2Cb2ac15a2344f7b7872c59d20633ad8b907586b12