reviewboard: implement automatic repository discovery (bug 1045736); r=smacleod
Currently, users must define a per-repository URL of the review
repository to push to. This can be slightly confusing as the names may
not map cleanly to what they are elsewhere. Also, the set of available
repositories may not be known.

This patch attempts to change that.

We implement functionality on the server that allows discovery of
available review repositories. Essentially, the server advertises that
it has a list of review repositories. When the client pushes to this
repository, it asks for the list of review repositories, figures out
which one (if any) is related, and automatically redirects to the
appropriate repository.

After this patch, clients can set a universal "review" path to the auto
review repository. `hg push review` will thus do the right thing (as
long as a review repo for the current repo is known, anyway).

A client repository is matched to a review repository by comparing the
root commit SHA-1.

This feature operates on the assumption that there is only one review
repository per root commit. This should be a fair assumption. If this
assumption ever breaks, there's a good chance we're doing something
wrong with the review repositories. That being said, the future may
require separate review repositories for e.g. security reviews. If this
happens, we can always introduce a new wire protocol command for
exposing the richer semantics needed to choose an appropriate
repository. In the mean time, YAGNI applies and simplicitly prevails.

--HG--
extra : rebase_source : 9b9ee25bea739646529bceb40d9070ad5c04802a
extra : amend_source : 7304fc42e6f186f70c2a02a55182af1050f8d1bf