ansible: improvements to reviewboard deployment
As discovered during a deployment, our rsync from src to www was not
doing the right thing because of rsync behavior on backslashes. Fix it.

As part of recovering from a production deployment hiccup, I discovered
that deploying the admin node from a fresh state wasn't working.
Creating some directories ensures we can remove directories and deploy
from a clean slate.