mozreview: Create attachments for all children (Bug 1055021). r=smacleod
An attachment is now created on Bugzilla for each child. No attachment is
created for the parent.

Review logic has changed a bit as well. For children, a ship-it results in an
r+ on the corresponding attachment from the reviewer. A review with no
ship-it clears any '?' review flag with the requestee being the reviewer, and
any '+' with the setter being the reviewer.

For parents, reviews with ship-its are not allowed. A review with no ship-it
results in a comment unassociated with any attachment, and no flag changes.

To avoid getting MozReview into a weird state, child requests are published
only after all attachments have been created in Bugzilla.

--HG--
extra : rebase_source : 14f5398bb36df994b31624f12d2bd215dc783276