mozreview: Prevent ship-its in the UI from parent review requests (Bug 1055021). r=smacleod
The required hook isn't yet available in Review Board, so this just uses
CSS rules to hide the ship-it options.  The .parent-request CSS class
also had to be moved from #review_request to the body, since the dialog
is a child of the body.  I moved .commit-request as well for consistency.

--HG--
extra : rebase_source : d2f45cfc9f54a24a61344554eafbfb90368dc734