ansible: install /etc/mercurial/hgrc on hgweb machines
Puppet recently stopped managing /etc/mercurial/hgrc. So take it over.

Since we don't have an hg-web role, create one. We'll eventually want to
move content from the standalone hgmo-extensions.yml file to roles.

The content of the hgrc file is equivalent to what is on the machines
today with the exception of the comment at the beginning of the file.

--HG--
extra : amend_source : b9dc3010e9879c1aeb1a644c6f3a1067c8061eae