reviewboard: Add expanded error messages for APIErrors when looking up reviewers (bug 1174884)
We expect APIErrors when looking up reviewers as that is what is returned if
a username does not exist. This adds code to check whether or not the error is
a 404 for an unknown user or a different error. If a different error is
encountered, the full error message is displayed.

--HG--
extra : commitid : 6miFyD30r5p
extra : rebase_source : ba3eff4d52fe83efffca90c794b6e40b044ba558