autoland: read default database dsn from config.json
Autoland accepts a --dsn argument to identify to which database it should
connect. At the moment it has a hard coded default value. This replaces that
with the database string found in config.json.

This also adds more globs to test-post-autoland-job.t to avoid races during the
tests.

--HG--
extra : commitid : Jej6HVm6GQi