mozreview: Always record cookies in session when logging in (bug 1177454). r=dminor
The Bugzilla authentication layer was not setting the user cookie variables,
which are saved to the session, when the login was performed with cookies.
This was probably left over from an early attempt to use Bugzilla's cookies
directly when authenticated with the UI.  Now we only use cookies with
API authentication, so we need to percolate them up to the middleware to save
in the session.

--HG--
extra : commitid : 2dmaUSydNEP
extra : rebase_source : 4d7e9d645052e4a32cff4b2f65eb084fd5bf380d