reviewboard: check capabilities on client; r=smacleod
The server now advertises a capability listing the required client
capabilities for interfacing with the server.

If clients see this capability, they will verify they meet the minimum
requirements to talk to the server and abort if not. Thus, all a server
needs to do to force all clients to upgrade is to advertise a new
required capability on the client.

--HG--
extra : commitid : 1dRzT59GK1h
extra : rebase_source : 4e719e774e5e847d96a03d64e604892c8fdf0de9